feat(): update 3.0.0

+++ bsin-targe-gateway-admin/src/main/java/org/apache/shenyu/admin/service/impl/PluginServiceImpl.java
@@ -20,6 +20,7 @@ package org.apache.shenyu.admin.service.impl;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.shenyu.admin.aspect.annotation.Pageable;
+import org.apache.shenyu.admin.mapper.NamespacePluginRelMapper;
 import org.apache.shenyu.admin.mapper.PluginMapper;
 import org.apache.shenyu.admin.model.dto.PluginDTO;
 import org.apache.shenyu.admin.model.entity.PluginDO;
@@ -28,21 +29,25 @@ import org.apache.shenyu.admin.model.page.CommonPager;
 import org.apache.shenyu.admin.model.page.PageResultUtils;
 import org.apache.shenyu.admin.model.query.PluginQuery;
 import org.apache.shenyu.admin.model.query.PluginQueryCondition;
+import org.apache.shenyu.admin.model.result.ConfigImportResult;
 import org.apache.shenyu.admin.model.vo.PluginSnapshotVO;
 import org.apache.shenyu.admin.model.vo.PluginVO;
+import org.apache.shenyu.admin.service.PluginHandleService;
 import org.apache.shenyu.admin.service.PluginService;
+import org.apache.shenyu.admin.service.configs.ConfigsImportContext;
 import org.apache.shenyu.admin.service.publish.PluginEventPublisher;
 import org.apache.shenyu.admin.transfer.PluginTransfer;
 import org.apache.shenyu.admin.utils.Assert;
-import org.apache.shenyu.common.constant.Constants;
-import org.apache.shenyu.common.utils.ListUtil;
 import org.apache.shenyu.admin.utils.SessionUtil;
 import org.apache.shenyu.admin.utils.ShenyuResultMessage;
 import org.apache.shenyu.common.constant.AdminConstants;
+import org.apache.shenyu.common.constant.Constants;
 import org.apache.shenyu.common.dto.PluginData;
 import org.apache.shenyu.common.exception.ShenyuException;
 import org.apache.shenyu.common.utils.JarDependencyUtils;
+import org.apache.shenyu.common.utils.ListUtil;
 import org.apache.shenyu.common.utils.LogUtils;
+import org.apache.shenyu.common.utils.UUIDUtils;
 import org.opengauss.util.Base64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -51,7 +56,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -65,23 +72,31 @@ public class PluginServiceImpl implements PluginService {
      * logger.
      */
     private static final Logger LOG = LoggerFactory.getLogger(PluginServiceImpl.class);
-    
+
     private final PluginMapper pluginMapper;
-    
+
     private final PluginEventPublisher pluginEventPublisher;
-    
+
+    private final PluginHandleService pluginHandleService;
+
+    private final NamespacePluginRelMapper namespacePluginRelMapper;
+
     public PluginServiceImpl(final PluginMapper pluginMapper,
-                             final PluginEventPublisher pluginEventPublisher) {
+                             final PluginEventPublisher pluginEventPublisher,
+                             final PluginHandleService pluginHandleService,
+                             final NamespacePluginRelMapper namespacePluginRelMapper) {
         this.pluginMapper = pluginMapper;
         this.pluginEventPublisher = pluginEventPublisher;
+        this.pluginHandleService = pluginHandleService;
+        this.namespacePluginRelMapper = namespacePluginRelMapper;
     }
-    
+
     @Override
     public List<PluginVO> searchByCondition(final PluginQueryCondition condition) {
         condition.init();
         return pluginMapper.searchByCondition(condition);
     }
-    
+
     /**
      * create or update plugin.
      *
@@ -119,6 +134,12 @@ public class PluginServiceImpl implements PluginService {
         if (CollectionUtils.isEmpty(plugins)) {
             return AdminConstants.SYS_PLUGIN_ID_NOT_EXIST;
         }
+        Optional<PluginDO> exist = plugins.stream()
+                .filter(value -> Objects.nonNull(this.namespacePluginRelMapper.selectByPluginId(value.getId())))
+                .findAny();
+        if (exist.isPresent()) {
+            return AdminConstants.NAMESPACE_PLUGIN_NOT_DELETE;
+        }
         // delete plugins.
         if (this.pluginMapper.deleteByIds(ListUtil.map(plugins, PluginDO::getId)) > 0) {
             // publish deleted event. synchronously delete and link data[selector,rule,condition,resource]
@@ -126,7 +147,7 @@ public class PluginServiceImpl implements PluginService {
         }
         return StringUtils.EMPTY;
     }
-    
+
     /**
      * plugin enabled.
      *
@@ -148,7 +169,7 @@ public class PluginServiceImpl implements PluginService {
         }
         return StringUtils.EMPTY;
     }
-    
+
     /**
      * find plugin by id.
      *
@@ -159,7 +180,7 @@ public class PluginServiceImpl implements PluginService {
     public PluginVO findById(final String id) {
         return PluginVO.buildPluginVO(pluginMapper.selectById(id));
     }
-    
+
     /**
      * find page of plugin by query.
      *
@@ -174,7 +195,7 @@ public class PluginServiceImpl implements PluginService {
                 .map(PluginVO::buildPluginVO)
                 .collect(Collectors.toList()));
     }
-    
+
     /**
      * query all plugin.
      *
@@ -186,17 +207,25 @@ public class PluginServiceImpl implements PluginService {
     }
     
     @Override
+    public List<PluginVO> listAllData() {
+        return pluginMapper.selectAll()
+                .stream()
+                .filter(Objects::nonNull)
+                .map(PluginVO::buildPluginVO).collect(Collectors.toList());
+    }
+    
+    @Override
     public List<PluginData> listAllNotInResource() {
         return ListUtil.map(pluginMapper.listAllNotInResource(), PluginTransfer.INSTANCE::mapToData);
     }
-    
+
     @Override
     public String selectIdByName(final String name) {
         PluginDO pluginDO = pluginMapper.selectByName(name);
         Objects.requireNonNull(pluginDO);
         return pluginDO.getId();
     }
-    
+
     /**
      * Find by name plugin do.
      *
@@ -209,13 +238,52 @@ public class PluginServiceImpl implements PluginService {
     }
 
     /**
-     *  activate plugin snapshot.
+     * activate plugin snapshot.
      *
      * @return List of plugins snapshot
      */
     @Override
     public List<PluginSnapshotVO> activePluginSnapshot() {
         return pluginMapper.activePluginSnapshot(SessionUtil.isAdmin() ? null : SessionUtil.visitor().getUserId());
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public ConfigImportResult importData(final List<PluginDTO> pluginList, final ConfigsImportContext context) {
+        if (CollectionUtils.isEmpty(pluginList)) {
+            return ConfigImportResult.success();
+        }
+        Map<String, PluginDO> existPluginMap = pluginMapper.selectAll()
+                .stream()
+                .filter(Objects::nonNull)
+                .collect(Collectors.toMap(PluginDO::getName, x -> x));
+        StringBuilder errorMsgBuilder = new StringBuilder();
+        int successCount = 0;
+        for (PluginDTO pluginDTO : pluginList) {
+            String pluginName = pluginDTO.getName();
+            // check plugin base info
+            if (existPluginMap.containsKey(pluginName)) {
+                errorMsgBuilder
+                        .append(pluginName)
+                        .append(",");
+                Optional.ofNullable(context).ifPresent(c -> c.getPluginTemplateIdMapping().put(pluginDTO.getId(), existPluginMap.get(pluginName).getId()));
+            } else {
+                String pluginId = UUIDUtils.getInstance().generateShortUuid();
+                Optional.ofNullable(context).ifPresent(c -> c.getPluginTemplateIdMapping().put(pluginDTO.getId(), pluginId));
+                pluginDTO.setId(pluginId);
+                PluginDO pluginDO = PluginDO.buildPluginDO(pluginDTO);
+                if (pluginMapper.insertSelective(pluginDO) > 0) {
+                    // publish create event. init plugin data
+                    successCount++;
+                }
+            }
+        }
+        if (StringUtils.isNotEmpty(errorMsgBuilder)) {
+            errorMsgBuilder.setLength(errorMsgBuilder.length() - 1);
+            return ConfigImportResult
+                    .fail(successCount, "import fail plugin: " + errorMsgBuilder);
+        }
+        return ConfigImportResult.success(successCount);
     }
     
     /**
@@ -228,8 +296,8 @@ public class PluginServiceImpl implements PluginService {
      * @see PluginCreatedEvent
      */
     private String create(final PluginDTO pluginDTO) {
-        Assert.isNull(pluginMapper.nameExisted(pluginDTO.getName()), AdminConstants.PLUGIN_NAME_IS_EXIST);
-        if (!Objects.isNull(pluginDTO.getFile())) {
+        Assert.isNull(pluginMapper.nameExisted(pluginDTO.getName()), "create" + AdminConstants.PLUGIN_NAME_IS_EXIST + pluginDTO.getName());
+        if (Objects.nonNull(pluginDTO.getFile())) {
             Assert.isTrue(checkFile(Base64.decode(pluginDTO.getFile())), AdminConstants.THE_PLUGIN_JAR_FILE_IS_NOT_CORRECT_OR_EXCEEDS_16_MB);
         }
         PluginDO pluginDO = PluginDO.buildPluginDO(pluginDTO);
@@ -248,8 +316,8 @@ public class PluginServiceImpl implements PluginService {
      * @return success is empty
      */
     private String update(final PluginDTO pluginDTO) {
-        Assert.isNull(pluginMapper.nameExistedExclude(pluginDTO.getName(), Collections.singletonList(pluginDTO.getId())), AdminConstants.PLUGIN_NAME_IS_EXIST);
-        if (!Objects.isNull(pluginDTO.getFile())) {
+        Assert.isNull(pluginMapper.nameExistedExclude(pluginDTO.getName(), Collections.singletonList(pluginDTO.getId())), AdminConstants.PLUGIN_NAME_IS_EXIST + pluginDTO.getName());
+        if (Objects.nonNull(pluginDTO.getFile())) {
             Assert.isTrue(checkFile(Base64.decode(pluginDTO.getFile())), AdminConstants.THE_PLUGIN_JAR_FILE_IS_NOT_CORRECT_OR_EXCEEDS_16_MB);
         }
         final PluginDO before = pluginMapper.selectById(pluginDTO.getId());
@@ -275,7 +343,7 @@ public class PluginServiceImpl implements PluginService {
                 return false;
             }
             Set<String> dependencyTree = JarDependencyUtils.getDependencyTree(file);
-            return dependencyTree.contains(AdminConstants.PLUGIN_ABSTRACR_PATH) || dependencyTree.contains(AdminConstants.PLUGIN_INTERFACE_PATH);
+            return dependencyTree.contains(AdminConstants.PLUGIN_ABSTRACT_PATH) || dependencyTree.contains(AdminConstants.PLUGIN_INTERFACE_PATH);
         } catch (Exception e) {
             LogUtils.error(LOG, "check plugin jar error:{}", e.getMessage());
             throw new ShenyuException(e);
